requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-auth